+++ lib/facebook/messenger/thread.rb
@@ -15,6 +15,10 @@ module Facebook
       module_function
 
       def set(settings, access_token:)
+        warn '[DEPRECATED] Facebook::Messenger::Thread and the thread settings API are deprecated as of Mar 02 2017.'\
+             ' Please use Facebook::Messenger::Profile instead.'
+        warn 'See https://developers.facebook.com/docs/messenger-platform/changelog for more information.'
+
         response = post '/thread_settings', body: settings.to_json, query: {
           access_token: access_token
         }
@@ -25,6 +29,10 @@ module Facebook
       end
 
       def unset(settings, access_token:)
+        warn '[DEPRECATED] Facebook::Messenger::Thread and the thread settings API are deprecated as of Mar 02 2017.'\
+             ' Please use Facebook::Messenger::Profile instead.'
+        warn 'See https://developers.facebook.com/docs/messenger-platform/changelog for more information.'
+
         response = delete '/thread_settings', body: settings.to_json, query: {
           access_token: access_token
         }